orders/flask/requirements.txt
APScheduler==3.9.1
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
binance-futures==0.0.1
binance-futures-connector==3.2.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
constantly==15.1.0
cryptography==38.0.1
Flask==2.2.2
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pytz==2022.2.1
requests==2.28.1
service-identity==21.1.0
six==1.16.0
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzlocal==2.1
urllib3==1.26.12
websocket-client==1.4.1
Werkzeug==2.2.2
zope.interface==5.4.0
